Clarification of the matrix assignment package
remove methods getNumberOfWorkersRows() and getNumberOfWorkersColumns() of the MatrixAssignment interface. They were only present for printing and they made no sense in that a partition can have distinct number of workers on different rows (for example in Modification3).

To allow the printing of a MatrixAssignment, the method toString() has been overloaded in both OptimalPartition and EquiMatrixAssignment.

The method verifyValidity() in Partition has been rename to valid() to be Java-friendly.

The parameter of the method getRegions has been changed from "isSRelation" to "dimension" to be consistent with the implemented interface.

Some comments have been added to the MatrixMapping package.