Modified Operator interface to produce streams
They used to produce a single tuple, but now they produce a list of
tuples. There is an abstract class called OneToOneOperator that reflects
the fact that many operators (all the ones already implemented) turn
given tuple into at most one new tuple. This class also makes it easier
to implement such operators and adapt old ones, as it uses the "one
tuple" version of the method to implement the "stream version"